5.1.2 Refactored click handlers. Render func moved to utils\render.js
