+++ src/main.js
@@ -9,7 +9,7 @@ import TripEventComponent from "./components/trip-event.js";
 import TripEventsComponent from "./components/trip-events.js";
 import {generateFilters} from "./mock/filters.js";
 import {generateEvents} from "./mock/trip-event.js";
-import {render, RenderPosition} from "./utils.js";
+import {render, RenderPosition} from "./utils/render.js";
 
 
 const EVENT_COUNT = 5;
@@ -17,10 +17,12 @@ const EVENT_COUNT = 5;
 const renderEvents = (tripEventsListElement, event) => {
   const onEditButtonClick = () => {
     tripEventsListElement.replaceChild(editFormComponent.getElement(), tripEventComponent.getElement());
+    document.addEventListener(`keydown`, onEscKeyDown);
   };
 
   const onEditFormClose = () => {
     tripEventsListElement.replaceChild(tripEventComponent.getElement(), editFormComponent.getElement());
+    document.removeEventListener(`keydown`, onEscKeyDown);
   };
 
   const onEscKeyDown = (evt) => {
@@ -28,38 +30,30 @@ const renderEvents = (tripEventsListElement, event) => {
 
     if (isEscKey) {
       tripEventsListElement.replaceChild(tripEventComponent.getElement(), editFormComponent.getElement());
+      document.removeEventListener(`keydown`, onEscKeyDown);
     }
   };
 
   const tripEventComponent = new TripEventComponent(event);
-  const editButton = tripEventComponent.getElement().querySelector(`button`);
-  editButton.addEventListener(`click`, () => {
-    onEditButtonClick();
-    document.addEventListener(`keydown`, onEscKeyDown);
-  });
+  tripEventComponent.setEditButtonClickHandler(onEditButtonClick);
 
   const editFormComponent = new EditFormComponent(event);
-  const editFormCloseButton = editFormComponent.getElement().querySelector(`.event__rollup-btn`);
+  editFormComponent.setCloseEditButtonClickHandler(onEditFormClose);
 
-  editFormCloseButton.addEventListener(`click`, () => {
-    onEditFormClose();
-    document.removeEventListener(`keydown`, onEscKeyDown);
-  });
-
-  render(tripEventsListElement, tripEventComponent.getElement(), RenderPosition.BEFOREEND);
+  render(tripEventsListElement, tripEventComponent, RenderPosition.BEFOREEND);
 };
 
 const renderTripEvents = (tripEventsComponent, events) => {
   let totalPrice = 0;
   if (events.length === 0) {
-    render(tripEventsComponent.getElement(), new NoEventsComponent().getElement(), RenderPosition.AFTERBEGIN);
+    render(tripEventsComponent.getElement(), new NoEventsComponent(), RenderPosition.AFTERBEGIN);
     return;
   }
 
-  render(tripEventsComponent.getElement(), new SortingComponent().getElement(), RenderPosition.BEFOREEND);
+  render(tripEventsComponent.getElement(), new SortingComponent(), RenderPosition.BEFOREEND);
 
   const tripDaysComponent = new TripDaysComponent();
-  render(tripEventsComponent.getElement(), tripDaysComponent.getElement(), RenderPosition.BEFOREEND);
+  render(tripEventsComponent.getElement(), tripDaysComponent, RenderPosition.BEFOREEND);
 
   const tripEventsListElement = tripDaysComponent.getElement().querySelector(`.trip-events__list`);
   events
@@ -74,15 +68,15 @@ const tripControls = tripMain.querySelector(`.trip-controls`);
 
 const filters = generateFilters();
 
-render(tripControls, new MenuComponent().getElement(), RenderPosition.AFTERBEGIN);
-render(tripControls, new FiltersComponent(filters).getElement(), RenderPosition.BEFOREEND);
+render(tripControls, new MenuComponent(), RenderPosition.AFTERBEGIN);
+render(tripControls, new FiltersComponent(filters), RenderPosition.BEFOREEND);
 
 const events = generateEvents(EVENT_COUNT);
 
 const main = document.querySelector(`.page-main`);
 const mainContainer = main.querySelector(`.page-body__container`);
 const tripEventsComponent = new TripEventsComponent();
-render(mainContainer, tripEventsComponent.getElement(), RenderPosition.BEFOREEND);
+render(mainContainer, tripEventsComponent, RenderPosition.BEFOREEND);
 renderTripEvents(tripEventsComponent, events);
 
-render(tripMain, new TripInfoComponent(0).getElement(), RenderPosition.AFTERBEGIN);
+render(tripMain, new TripInfoComponent(0), RenderPosition.AFTERBEGIN);
