+++ src/utils/render.js
@@ -1,0 +1,22 @@
+export const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
+
+export const render = (container, component, place) => {
+  switch (place) {
+    case RenderPosition.AFTERBEGIN:
+    container.prepend(component.getElement());
+    break;
+    case RenderPosition.BEFOREEND:
+    container.append(component.getElement());
+    break;
+  }
+};
+
+export const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+
+  return newElement.firstChild;
+};
